requirements.txt
asgiref==3.5.2
certifi==2024.2.2
cffi==1.15.1
Django==4.1
django-extensions==3.2.3
django-webpack-loader==2.0.1
djangorestframework==3.14.0
idna==3.4
iniconfig==1.1.1
markupsafe==2.1.3
packaging==23.1
pip==23.2.1
pluggy==1.0.0
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==23.2.0
pysocks==1.7.1
pytest==7.4.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
setuptools==68.0.0
sqlparse==0.4.4
tomli==2.0.1
Werkzeug==3.0.1
wheel==0.38.4

# Pip-only section (extracted from conda pip block)
async-timeout==4.0.3
cachetools==5.3.2
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
cryptography==41.0.4
deprecated==1.2.14
django-browser-reload==1.11.0
django-cors-headers==4.2.0
django-oauth-toolkit==2.3.0
django-storages==1.14.2
djangorestframework-simplejwt==5.3.0
google-api-core==2.16.1
google-auth==2.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
jwcrypto==1.5.0
msgpack==1.0.7
oauthlib==3.2.2
optional-django==0.1.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==1.7.1
pytest-django==4.8.0
redis==5.0.1
rsa==4.9
urllib3==2.0.5
wrapt==1.15.0
